Refactor Researcher page with new components
Extract functionality into new `ResearchPageContent` and `ResearchForm` components to simplify the `ResearcherPage` implementation. This change improves code maintainability by separating concerns and utilizing `requireAuth` for authentication checks.

Took 14 minutes
